# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CompanyResponse:
    properties:
      created_at:
        example: "2026-05-01T12:00:00Z"
        type: string
      description:
        example: A leading software company.
        type: string
      id:
        example: a1b2c3d4-...
        type: string
      industry:
        example: Technology
        type: string
      location:
        example: Colombo, Sri Lanka
        type: string
      name:
        example: TechCorp Lanka
        type: string
      website:
        example: https://techcorplanka.com
        type: string
    type: object
  dto.CreateCompanyRequest:
    properties:
      description:
        example: A leading software company.
        type: string
      industry:
        example: Technology
        maxLength: 100
        type: string
      location:
        example: Colombo, Sri Lanka
        maxLength: 150
        type: string
      name:
        example: TechCorp Lanka
        maxLength: 150
        minLength: 2
        type: string
      website:
        example: https://techcorplanka.com
        type: string
    required:
    - description
    - industry
    - location
    - name
    type: object
...
  dto.UpdateCompanyRequest:
    properties:
      description:
        example: Updated description.
        type: string
      industry:
        example: Technology
        maxLength: 100
        type: string
      location:
        example: Colombo 03, Sri Lanka
        maxLength: 150
        type: string
      name:
        example: TechCorp Lanka
        maxLength: 150
        minLength: 2
        type: string
      website:
        example: https://techcorplanka.com
        type: string
    type: object
...
  handler.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hireflow.com
    name: HireFlow Support
  description: A production-grade Job Recruitment REST API built with Go and Gin.
  license:
    name: MIT
  title: HireFlow API
  version: "1.0"
paths:
...
  /companies:
    get:
      description: Returns a list of all companies. Public endpoint.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CompanyResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: List all companies
      tags:
      - Companies
    post:
      consumes:
      - application/json
      description: Creates a new company. Admin only.
      parameters:
      - description: Company details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CompanyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                errors:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Create a company
      tags:
      - Companies
  /companies/{id}:
    delete:
      description: Soft-deletes a company. Admin only.
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a company
      tags:
      - Companies
    get:
      description: Returns a single company by ID. Public endpoint.
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CompanyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: Get a company
      tags:
      - Companies
    put:
      consumes:
      - application/json
      description: Updates company details. Recruiter only.
      parameters:
      - description: Company UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CompanyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a company
      tags:
      - Companies
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"